% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankeyNetwork.R
\name{WaterSankeyD3-shiny}
\alias{WaterSankeyD3-shiny}
\alias{sankeyNetworkOutput}
\alias{renderSankeyNetwork}
\title{Shiny bindings for WaterSankeyD3}
\usage{
sankeyNetworkOutput(outputId, width = "100\%", height = "500px")

renderSankeyNetwork(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{dimensions of the widget}

\item{expr}{expression to render}

\item{env}{environment}

\item{quoted}{logical, is expr quoted}
}
\description{
Output and render functions for using WaterSankeyD3 within Shiny applications.
}
